Tratamento de Exceptions

+++ Modulo_03/biblioteca/src/main/java/com/fuctura/biblioteca/services/CategoriaService.java
@@ -1,6 +1,9 @@
 package com.fuctura.biblioteca.services;
 
 import com.fuctura.biblioteca.dtos.CategoriaDTO;
+import com.fuctura.biblioteca.exceptions.DataIntegrityViolationException;
+import com.fuctura.biblioteca.exceptions.IllegalArgumentException;
+import com.fuctura.biblioteca.exceptions.ObjectNotFoundExeption;
 import com.fuctura.biblioteca.models.Categoria;
 import com.fuctura.biblioteca.repositories.CategoriaRepository;
 import org.modelmapper.ModelMapper;
@@ -9,7 +12,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.CancellationException;
+import java.util.stream.Collectors;
 
 @Service
 public class CategoriaService {
@@ -22,24 +25,44 @@ public class CategoriaService {
 
     public CategoriaDTO findById(Integer id) {
         Optional<Categoria> cat = categoriaRepository.findById(id);
-        return modelMapper.map(cat.orElse(null), CategoriaDTO.class);
+        if(cat.isPresent()){
+            return modelMapper.map(cat.orElse(null), CategoriaDTO.class);
+        }else {
+            throw new ObjectNotFoundExeption("Categoria não Encontrada");
+        }
     }
 
-    public List<Categoria> findAll() {
-        return categoriaRepository.findAll();
+    public List<CategoriaDTO> findAll() {
+        List<Categoria>lcat=categoriaRepository.findAll();
+        return lcat.stream().map(c -> modelMapper.map(c , CategoriaDTO.class)).collect(Collectors.toList());
     }
 
     public CategoriaDTO save(CategoriaDTO categoriaDTO) {
+        findByNome(categoriaDTO);
         Categoria cat = categoriaRepository.save(modelMapper.map(categoriaDTO, Categoria.class));
         return modelMapper.map(cat, CategoriaDTO.class);
     }
 
     public CategoriaDTO update(CategoriaDTO categoriaDTO) {
+        findById(categoriaDTO.getId());
+        findByNome(categoriaDTO);
         Categoria cat = categoriaRepository.save(modelMapper.map(categoriaDTO, Categoria.class));
         return modelMapper.map(cat, CategoriaDTO.class);
     }
 
     public void delete(Integer id){
+        findById(id);
+        Optional<Categoria> cat = categoriaRepository.findById(id);
+        if(!cat.get().getLivros().isEmpty()){
+            throw new DataIntegrityViolationException("Não é possivel remover a categoria por existir livros");
+        }
         categoriaRepository.deleteById(id);
+    }
+
+    private void findByNome(CategoriaDTO categoriaDTO){
+        Optional<Categoria> cat = categoriaRepository.findByNome(categoriaDTO.getNome());
+        if(cat.isPresent() && cat.get().getNome().equalsIgnoreCase(categoriaDTO.getNome())){
+            throw new IllegalArgumentException("Categoria já cadastrada");
+        }
     }
 }
